# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "onec",
    "version": "2.0.0"
  },
...
  "paths": {
...
    "/negotiation": {
      "post": {
        "operationId": "createDataNegotiation",
        "tags": [
          "DataNegotiations"
        ],
        "description": "Create a new DataNegotiation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "nullable": true
                  },
                  "client": {
                    "type": "string",
                    "nullable": true
                  },
                  "user": {
                    "type": "string",
                    "nullable": true
                  },
                  "tags": {
                    "type": "string",
                    "nullable": true
                  },
                  "step": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "nullable": true
                  },
                  "value": {
                    "type": "number",
                    "nullable": true
                  },
                  "partnerId": {
                    "type": "string",
                    "nullable": true
                  },
                  "startsDate": {
                    "type": "string",
                    "nullable": true
                  },
                  "observation": {
                    "type": "string",
                    "nullable": true
                  },
                  "averageGuide": {
                    "type": "number",
                    "nullable": true
                  }
                },
                "required": [
                  "title",
                  "client",
                  "user",
                  "tags",
                  "step",
                  "status",
                  "value",
                  "partnerId",
                  "startsDate",
                  "observation",
                  "averageGuide"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getNegotiation",
        "tags": [
          "negotiation"
        ],
        "description": "Get a list of Negotiation",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string",
                        "nullable": true
                      },
                      "client": {
                        "type": "string",
                        "nullable": true
                      },
                      "user": {
                        "type": "string",
                        "nullable": true
                      },
                      "tags": {
                        "type": "string",
                        "nullable": true
                      },
                      "step": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string"
                      },
                      "value": {
                        "type": "number",
                        "nullable": true
                      },
                      "startsDate": {
                        "type": "string",
                        "nullable": true
                      },
                      "observation": {
                        "type": "string",
                        "nullable": true
                      },
                      "averageGuide": {
                        "type": "number",
                        "nullable": true
                      },
                      "partnerId": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "nullable": true
                      },
                      "updatedAt": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "title",
                      "client",
                      "user",
                      "tags",
                      "step",
                      "status",
                      "value",
                      "startsDate",
                      "observation",
                      "averageGuide",
                      "partnerId",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/contestation/{id}": {
      "get": {
        "operationId": "getContestationById",
        "tags": [
          "contestation"
        ],
        "description": "Get a contestation by ID",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "product": {
                      "type": "string",
                      "nullable": true
                    },
                    "competence": {
                      "type": "string",
                      "nullable": true
                    },
                    "cnpj": {
                      "type": "string",
                      "nullable": true
                    },
                    "client": {
                      "type": "string",
                      "nullable": true
                    },
                    "percentage": {
                      "type": "number",
                      "nullable": true
                    },
                    "compensation": {
                      "type": "number",
                      "nullable": true
                    },
                    "honorary": {
                      "type": "number",
                      "nullable": true
                    },
                    "tax": {
                      "type": "number",
                      "nullable": true
                    },
                    "valueTj": {
                      "type": "number",
                      "nullable": true
                    },
                    "toPay": {
                      "type": "number",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "nullable": true
                    },
                    "observation": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "updatedAt": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "product",
                    "competence",
                    "cnpj",
                    "client",
                    "percentage",
                    "compensation",
                    "honorary",
                    "tax",
                    "valueTj",
                    "toPay",
                    "status",
                    "observation",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateContestation",
        "tags": [
          "contestation"
        ],
        "description": "Update a contestation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product": {
                    "type": "string",
                    "nullable": true
                  },
                  "competence": {
                    "type": "string",
                    "nullable": true
                  },
                  "cnpj": {
                    "type": "string",
                    "nullable": true
                  },
                  "client": {
                    "type": "string",
                    "nullable": true
                  },
                  "percentage": {
                    "type": "number",
                    "nullable": true
                  },
                  "compensation": {
                    "type": "number",
                    "nullable": true
                  },
                  "honorary": {
                    "type": "number",
                    "nullable": true
                  },
                  "tax": {
                    "type": "number",
                    "nullable": true
                  },
                  "valueTj": {
                    "type": "number",
                    "nullable": true
                  },
                  "toPay": {
                    "type": "number",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "nullable": true
                  },
                  "observation": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteContestation",
        "tags": [
          "contestation"
        ],
        "description": "Delete a contestation",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}